+++ src/tasks/helpers/index.ts
@@ -1,3 +1,3 @@
+export * from "../issue-token/helpers/issuer-settings-utils"
 export * from "./account-set-names"
-export * from "./issuer-settings-utils"
 export * from "./math"
