+++ src/tasks/issue-token/sub-tasks/index.ts
@@ -1,3 +1,4 @@
+export * from "./authorize-trustlines"
 export * from "./configure-issuer"
 export * from "./create-trustlines"
 export * from "./create-wallets"
